+++ src/controllers/users.controller.ts
@@ -86,62 +86,6 @@ export const updateUser = async (req: Request, res: Response) => {
 };
 
 /**
- * update profile
- * @method updateProfile
- * @memberof usersController
- * @param {object} req
- * @param {object} res
- * @returns {(function|object)} Function next() or JSON object
- */
-export const updateProfile = async (req: Request, res: Response) => {
-  const { _id } = req.params;
-  const { firstName, lastName, email, password } = req.body;
-
-  const foundUser = await User.findById({ _id });
-
-  if (!foundUser) {
-    return res.status(404).json({ status: 'failed', message: 'User does not exist' });
-  }
-
-  foundUser.firstName = firstName || foundUser.firstName;
-  foundUser.lastName = lastName || foundUser.lastName;
-  foundUser.email = email || foundUser.email;
-
-  if (password) {
-    foundUser.password = bcrypt.hashSync(password, 8);
-  }
-
-  const updatedUser = await foundUser.save();
-
-  const payload = {
-    _id: updatedUser._id,
-    email: updatedUser.email,
-    role: updatedUser.role.toLowerCase(),
-  };
-
-  const token = generateToken(payload, process.env.SECRET_KEY as string);
-
-  const data = {
-    _id: updatedUser._id,
-    firstName: updatedUser.firstName,
-    lastName: updatedUser.lastName,
-    email: updatedUser.email,
-    role: updatedUser.role.toLowerCase(),
-    token,
-  };
-
-  return res
-    .status(200)
-    .cookie('token', token, {
-      maxAge: 1000 * 60 * 60 * 24,
-      secure: false,
-      httpOnly: true,
-      // sameSite: 'lax',
-    })
-    .json({ status: 'success', data });
-};
-
-/**
  * delete user
  * @method deleteUser
  * @memberof usersController
