+++ src/routes/usersRoute.ts
@@ -1,15 +1,10 @@
 import { Router } from 'express';
-import {
-  getAllUsers,
-  getUser,
-  updateUser,
-  updateProfile,
-  deleteUser,
-} from '../controllers/users.controller';
+import { getAllUsers, getUser, updateUser, deleteUser } from '../controllers/users.controller';
 import { authorizedRoles, checkAuthToken } from '../middleware/authorization.middleware';
-import { validateUpdatedUser, validateUserProfile } from '../middleware/validation.middleware';
 import asyncHandler from '../middleware/asyncErrorHandler.middleware';
 import { roles } from '../utils/constants';
+import { validate } from '../middleware/validate.middleware';
+import { updateUserSchema } from '../schema/updateUserSchema';
 
 const router: Router = Router();
 
@@ -24,16 +19,10 @@ router.put(
   '/:_id',
   checkAuthToken,
   authorizedRoles([roles.ADMIN]),
-  validateUpdatedUser,
+  validate(updateUserSchema),
   asyncHandler(updateUser),
 );
-router.put(
-  '/:_id',
-  checkAuthToken,
-  authorizedRoles([roles.ADMIN, roles.SELLER, roles.CUSTOMER]),
-  validateUserProfile,
-  asyncHandler(updateProfile),
-);
+
 router.delete('/:_id', checkAuthToken, authorizedRoles([roles.ADMIN]), asyncHandler(deleteUser));
 
 export default router;
